+++ .storybook/main.js
@@ -1,0 +1,5 @@
+export default {
+  framework: '@storybook/html-vite',
+  stories: ['../stories/**/*.stories.js'],
+  core: { disableTelemetry: true },
+};